requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
astroid==2.15.6
backports.zoneinfo==0.2.1
black==23.7.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.6
colorama==0.4.6
coverage==7.2.7
cryptography==44.0.2
cssselect2==0.7.0
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.9
dj-database-url==2.0.0
dj-static==0.0.6
Django==4.2.16
django-allauth==0.61.0
django-cockroachdb==4.2
django-cors-headers==4.2.0
django-environ==0.10.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.16.1
flake8==5.0.4
google-ai-generativelanguage==0.1.0
google-api-core==2.24.2
google-auth==2.39.0
google-generativeai==0.1.0rc1
googleapis-common-protos==1.70.0
grpcio==1.70.0
grpcio-status==1.62.3
gunicorn==19.9.0
html5lib==1.1
idna==3.10
isort==5.12.0
isoweek==1.3.3
lazy-object-proxy==1.9.0
lxml==5.3.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.11.0.86
oscrypto==1.3.0
packaging==23.1
pathspec==0.11.1
pillow==10.4.0
pipenv==2024.4.1
platformdirs==3.9.1
proto-plus==1.26.1
protobuf==4.25.6
psycopg2==2.9.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.9.1
pycparser==2.22
pyflakes==2.5.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.8.0
pylint==2.17.4
pypdf==5.4.0
pypng==0.20220715.0
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.16.0
sqlparse==0.4.4
static3==0.7.0
svglib==1.5.1
tinycss2==1.4.0
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
virtualenv==20.30.0
webencodings==0.5.1
whitenoise==4.1.4
wrapt==1.15.0
xhtml2pdf==0.2.17
